src/agent/pyproject.toml
[project]
name = "sre-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore==1.3.1",
    "strands-agents==1.27.0",
    "strands-agents-tools==0.2.20",
    "mcp==1.26.0",
    "boto3==1.42.52",
    "httpx==0.28.1",
    "aws-opentelemetry-distro==0.15.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []
